+++ src/Services/Catalog/Product/Service/Product.php
@@ -25,7 +25,6 @@ use Bitrix24\SDK\Services\AbstractService;
 use Bitrix24\SDK\Services\Catalog\Common\ProductType;
 use Bitrix24\SDK\Services\Catalog\Product\Result\ProductResult;
 use Bitrix24\SDK\Services\Catalog\Product\Result\ProductsResult;
-
 use Psr\Log\LoggerInterface;
 
 #[ApiServiceMetadata(new Scope(['catalog']))]
@@ -71,7 +70,9 @@ class Product extends AbstractService
     )]
     public function add(array $productFields): ProductResult
     {
-        return new ProductResult($this->core->call('catalog.product.add', [
+        return new ProductResult($this->core->call(
+            'catalog.product.add',
+            [
                 'fields' => $productFields
             ]
         ));