+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Added
 
+- Added service `Services\Catalog\ProductImage` with support for `catalog.productImage.*` methods,
+  see [catalog.productImage.* methods](https://apidocs.bitrix24.com/api-reference/catalog/product-image/index.html) ([#537](https://github.com/bitrix24/b24phpsdk/issues/537)):
+    - `add` adds an image to a product, parent product, variation, or service, with batch calls support
+    - `get` gets information about a product image by its identifier
+    - `list` gets the list of images for a product, with batch calls support
+    - `delete` deletes a product image, with batch calls support
 - Added service `Services\Landing\Site\Service\Site` with support methods,
   see [landing.site.* methods](https://github.com/bitrix24/b24phpsdk/issues/267):
     - `add` adds a site